fix: coalesce ResizeObserver scrolls to prevent 1px visual flakiness
The useAutoScroll hook's ResizeObserver was setting scrollTop synchronously
on every resize event. During async content rendering (Shiki highlighting,
Mermaid diagrams, images), this caused multiple scrolls per frame with
slightly different scrollHeight values, leading to 1px scroll offset
flakiness in visual regression tests.

Changes:
- Add RAF coalescing to ResizeObserver callback in useAutoScroll
- Simplify story helpers to use a single RAF wait after message load
- Remove complex multi-frame scroll stabilization from stories
